Updated Version of the GNS3 topology
Created this file as a working copy to correct some issues I was seeing
while trying to use a saved copy of this GNS3 file that was modified by
another.  This file has the following changes from the Original:

- Automatically starts all nodes upon opening of the file
- Removed the private configuration references (it gave me errors when
starting this gns3 file that ultimately lead me to work on this in the
first place)
- Modified the file to keep configurations after you completely power
down the nodes.
- Renamed all the startup config references to match the startup config
files names saved